requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
frozenlist==1.3.3
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
magic-filter==1.0.9
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
Pillow==9.5.0
proglog==0.1.10
pydub==0.25.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
